+++ Unity/Assets/Scripts/Hotfix/Server/MengJing/State/StateTypeRemove_OnNotify.cs
@@ -14,7 +14,7 @@ namespace ET.Server
             if (stateComponent.IsStateBroadcastType(args.nowStateType))
             {
                 M2C_UnitStateUpdate.UnitId = unit.Id;
-                M2C_UnitStateUpdate.StateType = args.nowStateType;
+                M2C_UnitStateUpdate.StateType = (long)args.nowStateType;
                 M2C_UnitStateUpdate.StateValue = args.stateValue;
                 M2C_UnitStateUpdate.StateOperateType = 2;
                 M2C_UnitStateUpdate.StateTime = 0;
@@ -25,7 +25,7 @@ namespace ET.Server
                 if (unit.Type == UnitType.Player)
                 {
                     M2C_UnitStateUpdate.UnitId = unit.Id;
-                    M2C_UnitStateUpdate.StateType = args.nowStateType;
+                    M2C_UnitStateUpdate.StateType = (long)args.nowStateType;
                     M2C_UnitStateUpdate.StateValue = args.stateValue;
                     M2C_UnitStateUpdate.StateOperateType = 2;
                     M2C_UnitStateUpdate.StateTime = 0;
